fix(api): Minor Entity fixes
