+++ apps/api/src/product-types/dto/product-type-attribute.dto.ts
@@ -1,8 +1,19 @@
-import { OmitType } from '@nestjs/swagger';
-import { ProductTypeAttributeEntity } from '../entities';
+import { ApiProperty, OmitType } from '@nestjs/swagger';
+import { ProductTypeAttributeEntity, ProductTypeVariantAttributeEntity } from '../entities';
 
 export class ProductTypeAttributeDto extends ProductTypeAttributeEntity {
 }
 
 export class CreateProductTypeAttributeDto extends OmitType(ProductTypeAttributeDto, ['id']) {
+}
+
+export class ProductTypeVariantAttributeDto extends ProductTypeVariantAttributeEntity {
+}
+
+export class CreateProductTypeVariantAttributeDto extends OmitType(ProductTypeVariantAttributeDto, ['id', 'attribute', 'productType']) {
+	@ApiProperty()
+	attributeId: number;
+
+	@ApiProperty()
+	productTypeId: number;
 }
