+++ apps/api/src/products/dto/product-attribute-assignment.dto.ts
@@ -1,9 +1,7 @@
 import { OmitType, PartialType } from '@nestjs/swagger';
 import {
 	AssignedProductAttributeEntity,
-	AssignedProductAttributeValueEntity,
 	AssignedProductVariantAttributeEntity,
-	AssignedProductVariantAttributeValueEntity,
 } from '../entities/product-attribute-assignment.entity';
 
 // export class AssignedProductAttributeDto extends OmitType(AssignedProductAttributeEntity, []){
@@ -16,16 +14,6 @@ export class CreateAssignedProductAttributeDto extends OmitType(AssignedProductA
 export class UpdateAssignedProductAttributeDto extends PartialType(CreateAssignedProductAttributeDto) {
 }
 
-// export class AssignedProductAttributeValueDto extends OmitType(AssignedProductAttributeValueEntity, []){
-export class AssignedProductAttributeValueDto extends AssignedProductAttributeValueEntity {
-}
-
-export class CreateAssignedProductAttributeValueDto extends OmitType(AssignedProductAttributeValueDto, ['id']) {
-}
-
-export class UpdateAssignedProductAttributeValueDto extends PartialType(CreateAssignedProductAttributeValueDto) {
-}
-
 // export class AssignedProductVariantAttributeDto extends OmitType(AssignedProductVariantAttributeEntity, []){
 export class AssignedProductVariantAttributeDto extends AssignedProductVariantAttributeEntity {
 }
@@ -34,15 +22,5 @@ export class CreateAssignedProductVariantAttributeDto extends OmitType(AssignedP
 }
 
 export class UpdateAssignedProductVariantAttributeDto extends PartialType(CreateAssignedProductVariantAttributeDto) {
-}
-
-// export class AssignedProductVariantAttributeDto extends OmitType(AssignedProductVariantAttributeEntity, []){
-export class AssignedProductVariantAttributeValueDto extends AssignedProductVariantAttributeValueEntity {
-}
-
-export class CreateAssignedProductVariantAttributeValueDto extends OmitType(AssignedProductVariantAttributeValueDto, ['id']) {
-}
-
-export class UpdateAssignedProductVariantAttributeValueDto extends PartialType(CreateAssignedProductVariantAttributeValueDto) {
 }
