+++ apps/api/src/products/products.module.ts
@@ -1,21 +1,17 @@
 import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
 import {
 	AssignedProductAttributeEntity,
-	AssignedProductAttributeValueEntity,
 	AssignedProductVariantAttributeEntity,
-	AssignedProductVariantAttributeValueEntity,
 } from './entities/product-attribute-assignment.entity';
+import {
+	AssignedProductAttributeValueEntity,
+	AssignedProductVariantAttributeValueEntity,
+} from './entities/product-attribute-value-assignment.entity';
 import { ProductVariantEntity } from './entities/product-variant.entity';
 import { ProductEntity } from './entities/product.entity';
 import { ProductsController } from './products.controller';
 import { ProductsService } from './products.service';
-
-@Module({
-	controllers: [ProductsController],
-	providers: [ProductsService],
-})
-export class ProductsModule {
-}
 
 export const ProductsModuleEntities = [
 	ProductEntity,
@@ -25,3 +21,11 @@ export const ProductsModuleEntities = [
 	AssignedProductAttributeValueEntity,
 	AssignedProductVariantAttributeValueEntity,
 ];
+
+@Module({
+	imports: [TypeOrmModule.forFeature(ProductsModuleEntities)],
+	controllers: [ProductsController],
+	providers: [ProductsService],
+})
+export class ProductsModule {
+}
